Add script for cleaning up raw photos
Cameras that take raw images usually also save a jpg. Generally my
process is to use the jpg previews to figure out which ones to keep and
which to delete since they tend to load faster when viewing them.

After deleting all the jpg previews for photos I don't want, I use the
script added in this commit to also delete the corresponding raw
versions.