macOS: Remap Home and End to cmd+left and cmd+right
The default for macOS is that Home and End are document-based instead of
line-based. This is super annoying and hard to get used to.

This is fixed by remapping Home and End to the macOS equivalents of
cmd+left and cmd+right. There are exceptions for applications that
already do the right thing (terminals) as well as applications that
pass keys into non-macOS systems (RDP, VNC, VMs, etc).

To apply this mapping, install Karabiner-Elements, go to the "Complex
Modifications" pane, click "Add rule" and enable all the rules in the
"Home and End" tree.